Client side mana and health fix

Closes #36

See merge request rkilks/iti0301-2024-game!15

+++ core/src/ee/taltech/network/messages/UpdateMana.java
@@ -3,7 +3,7 @@ package ee.taltech.network.messages;
 public class UpdateMana {
 
     public Integer playerId;
-    public Integer mana;
+    public double mana;
 
     /**
      * Empty constructor for Kryonet.
